propysolver/requirements.txt
google-generativeai
google-api-python-client
anthropic
daphne
channels
redis